requirements.txt
streamlit
python-dotenv
cohere
google-genai
pinecone
sentence-transformers
langchain-text-splitters
pypdf
tiktoken
